refactor(path): split search path resolution internals
