+++ jpa-rsql-search-api/src/main/java/io/github/ggomarighetti/jparsqlsearch/path/ResolvedGenericArrayType.java
@@ -1,0 +1,11 @@
+package io.github.ggomarighetti.jparsqlsearch.path;
+
+import java.lang.reflect.GenericArrayType;
+import java.lang.reflect.Type;
+
+record ResolvedGenericArrayType(Type genericComponentType) implements GenericArrayType {
+    @Override
+    public Type getGenericComponentType() {
+        return genericComponentType;
+    }
+}
