+++ jpa-rsql-search-api/src/main/java/io/github/ggomarighetti/jparsqlsearch/path/ResolvedPath.java
@@ -1,0 +1,12 @@
+package io.github.ggomarighetti.jparsqlsearch.path;
+
+import java.util.Set;
+
+record ResolvedPath(
+        Class<?> type,
+        boolean traversesCollection,
+        boolean collectionValued,
+        int depth,
+        Set<String> joinedPaths,
+        Set<String> toManyPaths) {
+}
